Username: jasplin Fixes:    Compile on Solaris RevBy:	  Marius Bugge Monsen Details:  Library functions ceilf(), logf(), and strtold() 	  not available on Solaris.
[git-p4: depot-paths = "//depot/qt/main/": change = 249929]

+++ src/gui/graphicsview/qgraphicsscene.cpp
@@ -332,7 +332,7 @@ void QGraphicsScenePrivate::resetIndex()
 
 static inline int intmaxlog(int n)
 {
-    return  (n > 0 ? qMax(int(::ceilf(::logf(float(n)) / ::logf(float(2)))), 5) : 0);
+    return  (n > 0 ? qMax(int(::ceil(::log(double(n)) / ::log(double(2)))), 5) : 0);
 }
 
 /*!
